full support for bitwise floats

+++ src/cxx/chaoscore/base/data/BitwiseFloat.hpp
@@ -12,14 +12,24 @@ namespace data
  * \brief Object that can be used to read and write the bits of a floating point
  *        number.
  *
- * TODO: DOC
+ * Floating point numbers are composed of three sections: sign, exponent, and
+ * mantissa. These sections are laid out like so:
+ *
+ * \code
+ * 0 00000000 00000000000000000000000
+ * ^ \------/ \---------------------/
+ * |     |                |
+ * \ 8-bit exponent       |
+ *  \                23-bit mantissa
+ * sign bit
+ * \endcode
+ *
+ * This object provides functions for reading and writing these sections of a
+ * floating point number.
  */
 union BitwiseFloat
 {
 public:
-
-    // TODO: getters and setters
-
 
     //--------------------------------------------------------------------------
     //                             PUBLIC ATTRIBUTES
@@ -48,27 +58,38 @@ public:
     //--------------------------------------------------------------------------
 
     /*!
-     * \brief Retrieves the sign bit of the floating point number.
+     * \brief Retrieves the sign bit of this floating point number.
      */
     bool get_sign_bit() const;
 
-    // TODO: DOC
+    /*!
+     * \brief Sets the sign bit (most significant bit) of this float.
+     */
     void set_sign_bit( bool sign );
 
     /*!
-     * \brief Retrieves the exponent section of the floating point number.
+     * \brief Retrieves the 8-bit exponent section of this floating point
+     *        number.
      */
     chaos::uint32 get_exponent() const;
 
-    // TODO: DOC
-    void set_exponent( chaos::uint32 exponent );
+    /*!
+     * \brief Sets the 8-bit exponent section of this floating point number.
+     */
+    void set_exponent( chaos::uint8 exponent );
 
     /*!
-     * \brief Retrieves the mantissa section of the floating point number.
+     * \brief Retrieves the 23-bit mantissa section of this floating point
+     *        number.
      */
     chaos::uint32 get_mantissa() const;
 
-    // TODO: DOC
+    /*!
+     * \brief Sets the 23-bit mantissa section of this floating point number.
+     *
+     * \note While a 32-bit input value is accepted only the least significant
+     *       23-bits of this value will be used.
+     */
     void set_mantissa( chaos::uint32 mantissa );
 };
